refactor tests and update changelog
